I've updated the app with the following:
Unlock key - the sound will occur immediately when the turn is completed and it now counts the repetitions.

Slow Pour - the activity will end 15% sooner so they will not need to pour as far.

Activity Selection - There was a bug where first choosing a Mug activity and then choosing a non-mug activity would still make the app go to the 'put in mug screen'. There was also one where choosing a different activity after selecting a hand would not override the previous choice.

Screen Lock / Orientation - The screen throughout the application will never turn off now or change from portrait programmatically (so we wont need to play with phone settings).

Post Workout - The order that the Text to speech voice talks aligns with the order the grades are presented.

I'm still getting the quick turn more responsive. For now, I can meet you on South Campus to install the update or email you the file.